requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
python-dotenv==1.0.0
httpx
lxml==4.9.3
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.4.2
ollama
sqlalchemy==2.0.22
selenium
webdriver-manager
schedule
cloudscraper
flask
fake-useragent
python-multipart
bs4
fake_useragent
feedparser
python-dateutil